backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
redis==5.2.0

# Google APIs
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.153.0

# AI Providers
anthropic==0.39.0
openai==1.54.5

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Utilities
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytz==2024.2

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
